reverse_proxy_api/requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0; python_version >= '3.8'
anyio==3.7.1; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
chardet==5.2.0; python_version >= '3.7'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cryptography==41.0.5
dynaconf==3.2.3; python_version >= '3.8'
fastapi==0.103.2; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httptools==0.6.1
idna==3.4; python_version >= '3.5'
jsonschema==4.20.0; python_version >= '3.8'
jsonschema-path==0.3.2; python_full_version >= '3.8.0' and python_full_version < '4.0.0'
jsonschema-specifications==2023.11.1; python_version >= '3.8'
lazy-object-proxy==1.9.0; python_version >= '3.7'
openapi-schema-validator==0.6.2; python_full_version >= '3.8.0' and python_full_version < '4.0.0'
openapi-spec-validator==0.7.1; python_full_version >= '3.8.0' and python_full_version < '4.0.0'
packaging==23.2; python_version >= '3.7'
pathable==0.4.3; python_full_version >= '3.7.0' and python_full_version < '4.0.0'
prance==23.6.21.0; python_version >= '3.8'
prometheus-client==0.17.1; python_version >= '3.6'
pycparser==2.21
pydantic==2.5.1; python_version >= '3.7'
pydantic-core==2.14.3; python_version >= '3.7'
pyjwt[crypto]==2.8.0; python_version >= '3.7'
python-dotenv==1.0.0
pyyaml==6.0.1
referencing==0.31.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
rfc3339-validator==0.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rpds-py==0.13.1; python_version >= '3.8'
ruamel.yaml==0.18.5; python_version >= '3.7'
ruamel.yaml.clib==0.2.8; python_version < '3.13' and platform_python_implementation == 'CPython'
simplejson==3.19.2; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
starlette==0.27.0; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
urllib3==2.1.0; python_version >= '3.8'
uvicorn[standard]==0.23.2; python_version >= '3.8'
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
